refactor(tree): extract shim detection module for tree advisor
